avoid changing directory in shell script
Since chee uses current working directory to determine repository mode,
it must not be changed by `chee` shell script.

When building the final jar, the classpath is included in the MANIFEST
file, so that it can be run if it is put into a directory containing the
dependency jars. The `chee` shell script then only needs to run this
jar.

This should fix #4.